Added some basic CLI tests
Some base tests copied from other repo to get the cli tests up and running.
Basic tests for add customer and add user.
Added new fixture 'disk_store' that is supposed to be a temporary db on disk that the cli can use (could not get it to work with un-named db).