Change read count type on statistics model (#2093) (patch)
Description
This PR fixes an issue discovered with the statistics table discovered when populating it with old data. Some of the read counts exceed the upper bound of an Integer (2147483647) in Python. Does it make sense to have read counts bigger than this? Maybe the issue is not with the model but with invalid data.