fix(bcl converter fetching for demux post postprocessing api) (#2375) (minor)
# Description
This PR removes the functionality of the demux post processing api to determine the bcl converter based on the directory structure. The approach would fail for instances where old flow cells were demultiplexed with bcl2fastq for which the folder structure is incomplete. Instead the bcl converter is determined by the sequencer type.

This PR also sets the default option of "@OPTION.BCL_CONVERTER" to none. This change is necessary to allow for automated setting of bcl converter options but to allow for manual overrides if required (i.e. demultiplexing a novaseq6000 flow cell using bcl2fastq).

# Changed
removed tests and logic around is_bcl2fastq_demux_folder_structure
sets the default option of @OPTION.BCL_CONVERTER to none.