fix(fetch HK sample sheet for validation new demux post processing flow) (#2381) (patch)
## Description

This PR changes the location of the sample sheet used in validation during the new demultuplexing post processing flow.
With these changes the sample sheet location is in housekeeper and not the `flow_cells/flow_cell_dir_name`. 

These changes go paired with that upon sample sheet creation for a given flow cell, the sample sheet is directly stored in housekeeper.