Remove strict qc controls for uploading microsalt cases to HK (#2505)(patch)
In #1655 a qc control was implemented for microSALT cases, which caused some complete cases not to update their status in statusDB from "running". This caused the files never to be uploaded to Housekeeper. As suggested in #2500, the qc controls were removed, and all the complete cases are returned when `get_cases_to_store` is called. The function `get_cases_to_store` in MicrosaltAPI, which overrode the class from AnalysisAPI, was removed, using the class from the base class with the intended behaviour.

### Added
- A test function for `get_cases_to_store ` when called through the MicrosaltAPI


### Fixed

- Removed functions `is_qc_required`, `get_completed_cases`, and `get_cases_to_store `. This last one is implemented in the base class with the desired functionality
- Remove tests for old `get_cases_to_store`