Make hamming distance calculation different for i7 and i5 (#2867)(patch)
## Description
Addresses #2857, partial rollback of #2683. 
Changes the way in which Hamming distance is calculated for index 2 (i5). If the override cycles of the sample sheet are reverse complemented, it uses the tails of the indexes to calculate the Hamming distance. It the override cycles value is not reverse complemented, the function uses the head of the indexes. 

If, for example, we had two i5 on the same lane with sequences CGGAACTG and CAGAACAGAA, the heads hamming distance would be calculated

```
CGGAACTG
|X||||X|      ---> Hamming distance = 2
CAGAACAGAA
```

While the tail comparison would be

```
  CGGAACTG
  XXXX|XXX   ---> Hamming distance = 7
CAGAACAGAA
```

This table summarises the new rule

| **Reverse complement** 	| **Override cycles example** 	| **String comparison** 	| **Barcode mismatch value<br> for the sequences above** 	|
|------------------------	|-----------------------------	|-----------------------	|----------------------------	|
| True                   	| `Y151;I8N2;N2I8;Y151`       	| Tails                 	| 1                          	|
| False                  	| `Y151;I8N2;I8N2;Y151`       	| Heads                 	| 0                          	|


### Added

- Function to calculate barcode mismatches for index 2
- Test for new function

### Changed

- Renamed previous hamming distance for indexes to only index 
- Parametrised tests