feat - include index settings in sample sheet header (#2931)(patch)
## Description
We need to implement a new validation for sample sheets that checks if the override cycles are correct. As in many cases we don't have the run parameters file to check the correct configuration of the cycles, this PR adds a field to the v2 sample sheet with the index settings used. This is enough to determine the correct override cycles for each sample.
Note: the validation will be done in another PR

### Added

- A new entry to the Header section of the sample sheet with the name of the index settings used