Remove alternate patterns to find fastq in flow cells (#3160)(patch)
This PR introduces a very small change in the codebase, and a somehow bigger refactoring of the post-processing tests.

The change in the codebase closes #3129, removing the flow cell folder structure patterns involving a "Unaligned" directory, which is deprecated and leaving only two patterns: The flat structure, which has all the fastq files in the demux directory per flow cell, and the pattern for flow cells demultiplexed on the sequencer. This change is entirely in the file `cg/meta/demultiplex/utils.py`.

The refactoring of the tests fixes all the tests that the codebase change broke and adds testing of the remaining functionalities that is **coherent** with the sample sheet and demultiplex testing done before, i.e. using only the same _canonical_ flow cells.

### Added

- A fixture environment consisting of a store, a context and a Post-processing API based on the seven canonical flow cells for demultiplexing (2 HiSeqX, 2HiSeq2500, 2 NovaSeq6000 and 1 NovaSeqX) The new fixtures are:
  - `updated_store_with_demultiplexed_samples`
  - `updated_demultiplex_context`
  - `updated_demultiplex_context`
  - `seven_canonical_flow_cells`
  - `seven_canonical_flow_cells_selected_sample_ids`
- Files for the new demultiplexed flow cell fixtures
### Changed

- Removed all patterns involving `Unaligned` in `cg/meta/demultiplex/utils.py`
- Moved some fixtures to fixture pluggins