change external data API (#2892)(minor)
## Description
Fixes #2864. The addition of external samples now also includes the samples to the Housekeeper bundles. Also, the operations of adding and including to Housekeeper and starting cases are independent per sample. In other words, before this PR we aborted the procedure if one of the samples had a problem, now only the problematic sample is aborted and the rest are processed as normal, see changes in function `ExternalAPI.add_transfer_to_housekeeper`.


### Added

- New functions in `ExternalAPI` and their tests

### Changed

- Modified the `ExternalAPI` to be more modular and readable, renaming internal functions with preceding `_`.
- Made `dry_run`, `ticket`, `customer` and `force` attributes fo the class so that they are not passed around all intermediate functions.
- Moved static functions to more appropriate modules (md5 function to checksum and `are_all_fastq_valid` to utils

### Fixed

- Method `ExternalAPI.add_transfer_to_housekeeper` so that it includes files to bundles and process every sample independently

---------

Co-authored-by: Henrik Stranneheim <henrik.stranneheim@scilifelab.se>