Parse metrics from smrtlnk datasets file (#3436)(patch)
## Description
Closes https://github.com/Clinical-Genomics/add-new-tech/issues/63
Parses the final metrics for PacBio. This last metrics file has a structure different from the previous ones. The desired metrics are not inside an `attributes` section in the json, but in the only existing section. Parsing for the other metrics had to be modified to account for the parsing of this file and make it more DRY.

### Added

- Metrics model
- Function to parse specifically the smrtlink-datasets file in the metrics parser
- Constants for metrics names
- Fixture for parsed metrics and fixture file

### Changed

- Merged `_parse_report` and `parse_attributes_to_model` methods of the metrics parser into one, called `parse_report_to_model`.
- Updated tests of these functions

### Fixed

- Implemented `TypeVar` from the `typing` module to better type-hint the different metrics classes, so just typing the parent class (BaseModel from Pydantic)