Set PacBio sequencing times (#3518)
## Description
Closes https://github.com/Clinical-Genomics/add-new-tech/issues/88

This PR parses the datetime for the start of the sequencing and the completion of the sequencing. The dataset metrics hold a proxy for the sequencing completed at date in a variable called `createdAt`. The sequencing started at can be extracted from the movie name, in any of the file names.