removing old sample and order models pt 1 (#4097)
## Description
The old validation used pydantic models such as `OrderIn`, and samples for each ordertype that new ones have replaced, therefore they need to be deleted. They are still called in many tests set to fail. The tests are still kept as they need to be translated into the new validation. This PR attempts to remove the pydantic models and fix the tests. Many changes are involved, so I separated the PR with an incomplete working set of changes.

### Added

- Fixtures for taxprofiler order
- Store fixture for order integration tests (includes all application and a maf order. To be expanded with organisms and everything needed for the tests)

### Changed

- Fixed integration tests for Balsamic, Fastq, Fuffy, and Metagenome

### Fixed

- Removed completely unused models
---------

Co-authored-by: Isak Ohlsson Ångnell <40887124+islean@users.noreply.github.com>