add error handling to fastq clean (#4206)(patch)
## Description
Closes #3817 
Adds error handling to the CLI command `cg compress clean fastq` so that any exception raised when cleaning a single sample is caught and logged without affecting the cleaning of other samples. Additionally, adds a check that verifies that a sample is not part of a newer case before cleaning.

### Added

- Error handling

### Changed
- The way samples to clean are fetched, now we verify if they belong to a case that is not cleanable.

Co-authored-by: eliott <eliottob6@gmail.com>
Co-authored-by: EliottBo <112384714+eliottBo@users.noreply.github.com>