Fix(dragen demux) Read bashrc to source the correct path
## Description
With the latest dragen version (4.4.4) which is running on cg-dragen2. The dragen installation does not add the dragen binary to the bath.

I changed the bashrc for hiseq.clinical to export it but Slurm does not load the bashrc per default. This PR adds a flag to make it source it

### Added

- Added the -l flag to the Sbatch header to run it as a login-session and thereby loading the bashrc