Block retrieval of SPRING files if any file is still archiving (#4616)(patch)
## Description
The retrieval of SPRING files behaves unexpectedly when their archiving takes too long. Specifically, when retrieving files from a sample that is currently being archived, it results in unexpected behaviour, as documented in https://github.com/Clinical-Genomics/bug-reports/issues/42. For this reason, this PR implements a check that stops a retrieval process of a sample if any of its files are currently being archived, raising a `SampleFilesCurrentlyArchivingError`.

### Added

- Support for checking if any of a sample's files is being archived, and raise an error if that is the case.
- Test for this new functionality

### Changed

- Modified existing tests to account for new functionality

### Fixed

- Removed unused fixtures