Fix(Panels) Change name of inherited cancer panel (#4792)
## Description
The panel with the abbreviation `"Inherited cancer"` will change name to IHC since the space in the abbreviation causes problems. Since we have a hard-coded constant in the code for which panels to always include we need to change it.

I've also copied the "Inherited cancer" panel in the database and named it IHC.

### Fixed

- Replaced "Inherited cancer" in the panel master list, with the new abbreviation